Check for errors in identify()
The Special:OAuth/identify endpoint returns a '200 OK' HTTP status code
even if the API request itself failed for some reason. The JSON response
payload can be checked by looking for an 'error' member in the response
body. This can be very help when debugging configuration issues such as
attempting to verify identity against a wiki where the consumer token is
invalid.